Completed integration of delete social club

+++ src/DeleteSocialClub/deleteSocialClub.component.ts
@@ -1,23 +1,34 @@
 import { Component, Input } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { ActivatedRoute, Router } from '@angular/router';
+import { ActivatedRoute, Router, RouterModule } from '@angular/router';
+import { FormBuilder, FormGroup, ReactiveFormsModule, Validators } from '@angular/forms';
 
 @Component({
   selector: 'app-delete-social-club',
   standalone: true,
-  imports: [CommonModule],
+  imports: [CommonModule, ReactiveFormsModule, RouterModule],
   templateUrl: './deleteSocialClub.component.html',
   styleUrl: './deleteSocialClub.component.css',
 })
 export class DeleteSocialClubComponent {
   @Input() clubName: string | undefined;
 
+  deleteForm: FormGroup;
   clubID = '';
+  formName = '';
+  isAPILoading = false;
+  showsuccessToast = false;
+  showfailToast = false;
   constructor(
     private router: Router,
-    private route: ActivatedRoute
+    private route: ActivatedRoute,
+    private fb: FormBuilder
   )
-  {}
+  {
+    this.deleteForm = this.fb.group({
+      name: ['', Validators.required]
+    });
+  }
 
   ngOnInit(): void {
     this.route.params.subscribe(params => {   // Get the event ID from the URL
@@ -36,30 +47,59 @@ export class DeleteSocialClubComponent {
   deleteClub() {
     this.route.params.subscribe(params => {   // Get the event ID from the URL
       this.clubID = params['id'];
-
-      try {
-        fetch('https://events-system-back.wn.r.appspot.com/api/socialclubs/' + this.clubID, {
-          method: 'DELETE',
-          credentials: 'include',
-          headers: {
-            'Content-Type': 'application/json'
-          },
-          body: '{}'
-        })
-        .then(response => {
-          return response.json();
-        })
-        .then(data => {
-          console.log(data);
-          console.log("Delete successful");
-        })
-      }
-      catch (error) {
-        console.error('Error:', error);
-        console.error('Error during deletion:', error);
-      }
       
-      this.router.navigate(['/socialclublisting']);
+      if(this.deleteForm.valid) {
+        this.isAPILoading = true;
+        this.formName = this.deleteForm.get('name')?.value
+        
+        if(this.formName.toLowerCase() === this.clubName?.toLowerCase()) {
+          fetch('https://events-system-back.wn.r.appspot.com/api/socialclubs/' + this.clubID, {
+            method: 'DELETE',
+            credentials: 'include',
+            headers: {
+              'Content-Type': 'application/json'
+            },
+            body: '{}'
+          })
+          .then(data => {
+            this.showsuccessToast = true;
+
+            setTimeout(() => {
+              this.showsuccessToast = false;
+              window.history.back();
+              // this.router.navigate(['/socialclublisting']);
+            }, 5000);
+          })
+          .catch ((error) => {
+            this.showfailToast = true;
+            this.isAPILoading = false;
+
+            setTimeout(() => {
+              this.showfailToast = false;
+            }, 10000);
+            console.error('Error:', error);
+          });
+        }
+        else
+        {
+          this.showfailToast = true;
+          this.isAPILoading = false;
+
+          setTimeout(() => {
+            this.showfailToast = false;
+          }, 10000);
+          console.error("Incorrect club name: " + this.clubName?.toLowerCase())
+        }
+      }
+      else {
+        this.showfailToast = true;
+        this.isAPILoading = false;
+
+        setTimeout(() => {
+          this.showfailToast = false;
+        }, 10000);
+        console.error("Please enter a name")
+      }
     });
   }
 }
